server/pyproject.toml
[project]
name = "eina-server"
version = "0.1.0"
description = "Eina server"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dependency-injector>=4.44.0",
    "fastapi>=0.115.0",
    "graphene>=3.4.3",
    "gunicorn>=23.0.0",
    "sqlalchemy>=2.0.36",
    "starlette-graphene3>=0.6.0",
    "strawberry-graphql>=0.253.1",
    "uvicorn>=0.32.1",
]

[tool.uv]
dev-dependencies = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.3",
    "ruff>=0.7.4",
]
